requirements-e2e.txt
# E2E Testing Dependencies (not included in production)
playwright>=1.40.0
pytest-playwright>=0.5.0
